refactored payment, reservation, and reviews
